Refactor cloud_proxy into storage
